% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findGenes}
\alias{findGenes}
\title{Runs prodigal on AAStringSet}
\usage{
findGenes(genome, quiet = TRUE)
}
\arguments{
\item{genome}{A DNAStringSet to search in (usually a full genome)}
}
\value{
An AAStringSet with the proteins detected
}
\description{
Runs prodigal on AAStringSet
}

